refactor(provider): collapse ProviderProfile to schema-as-struct
Flatten the ProviderProfile aggregate root — the last un-flattened Provider
entity — from DDD/Ports & Adapters to conventional Phoenix. Fuse the domain
model, Ecto schema, mapper, repository, ChangeProviderProfile adapter, the five
provider use-cases and the queries module into one schema-as-struct at the
context root, with commands/queries inlined onto KlassHero.Provider.

- profile_status: plain :string + validate_inclusion -> Ecto.Enum
  [:draft, :active], collapsing the three string<->atom mapping sites
- Frozen public API: {:validation_error, _}, :duplicate_resource,
  :already_active and :not_found error shapes preserved (facade tests pin them)
- Repointed the facade's own two joins, cross-context accounts/user.ex has_one,
  provider_events, the parked belongs_to aliases, presenters, LiveViews, the
  Backpex admin resource and check_provider_verification_status onto the facade
- Dropped the for_storing/for_querying_provider_profiles DI keys and
  ProviderProfileRepository; forms pin as: :provider_profile_schema
- Reworked the now-mapper-less RepositoryHelpers test with a local identity mapper
