+++ config/config.exs
@@ -18,7 +18,6 @@ alias KlassHero.Messaging.Adapters.Driving.Events.MessagingEventHandler
 alias KlassHero.Messaging.Adapters.Driving.Workers.MessageCleanupWorker
 alias KlassHero.Messaging.Adapters.Driving.Workers.RetentionPolicyWorker
 alias KlassHero.Provider.Adapters.Driven.ACL.ParticipationSessionStatsACL
-alias KlassHero.Provider.Adapters.Driven.Persistence.Repositories.ProviderProfileRepository
 alias KlassHero.Provider.Adapters.Driven.Persistence.Repositories.ProviderProgramRepository
 alias KlassHero.Provider.Adapters.Driven.Persistence.Repositories.SessionDetailsRepository
 alias KlassHero.Provider.Adapters.Driven.Persistence.Repositories.SessionStatsRepository
@@ -193,8 +192,6 @@ config :klass_hero, :messaging,
 # Configure Provider bounded context
 config :klass_hero, :provider,
   repo: KlassHero.Repo,
-  for_storing_provider_profiles: ProviderProfileRepository,
-  for_querying_provider_profiles: ProviderProfileRepository,
   for_querying_session_details: SessionDetailsRepository,
   for_querying_session_stats: SessionStatsRepository,
   for_resolving_session_stats: ParticipationSessionStatsACL,
