+++ test/klass_hero/provider/application/commands/providers/provider_verification_test.exs
@@ -13,8 +13,7 @@ defmodule KlassHero.Provider.Application.Commands.Providers.ProviderVerification
   import KlassHero.EventTestHelper
 
   alias KlassHero.AccountsFixtures
-  alias KlassHero.Provider.Application.Commands.Providers.UnverifyProvider
-  alias KlassHero.Provider.Application.Commands.Providers.VerifyProvider
+  alias KlassHero.Provider
   alias KlassHero.ProviderFixtures
 
   setup do
@@ -28,7 +27,7 @@ defmodule KlassHero.Provider.Application.Commands.Providers.ProviderVerification
     test "sets provider as verified", %{provider: provider, admin: admin} do
       params = %{provider_id: provider.id, admin_id: admin.id}
 
-      assert {:ok, verified} = VerifyProvider.execute(params)
+      assert {:ok, verified} = Provider.verify_provider(params.provider_id, params.admin_id)
 
       assert verified.verified == true
       assert verified.verified_at != nil
@@ -37,7 +36,7 @@ defmodule KlassHero.Provider.Application.Commands.Providers.ProviderVerification
     test "sets verified_at timestamp", %{provider: provider, admin: admin} do
       params = %{provider_id: provider.id, admin_id: admin.id}
 
-      {:ok, verified} = VerifyProvider.execute(params)
+      {:ok, verified} = Provider.verify_provider(params.provider_id, params.admin_id)
 
       # Verify the timestamp is set and is a recent DateTime
       assert %DateTime{} = verified.verified_at
@@ -48,7 +47,7 @@ defmodule KlassHero.Provider.Application.Commands.Providers.ProviderVerification
 
     test "publishes integration event", %{provider: provider, admin: admin} do
       params = %{provider_id: provider.id, admin_id: admin.id}
-      {:ok, _} = VerifyProvider.execute(params)
+      {:ok, _} = Provider.verify_provider(params.provider_id, params.admin_id)
 
       event = assert_integration_event_published(:provider_verified)
       assert event.entity_id == provider.id
@@ -59,7 +58,7 @@ defmodule KlassHero.Provider.Application.Commands.Providers.ProviderVerification
     test "returns error when provider not found", %{admin: admin} do
       params = %{provider_id: Ecto.UUID.generate(), admin_id: admin.id}
 
-      assert {:error, :not_found} = VerifyProvider.execute(params)
+      assert {:error, :not_found} = Provider.verify_provider(params.provider_id, params.admin_id)
     end
 
     test "is idempotent - verifying already verified provider succeeds", %{
@@ -69,11 +68,11 @@ defmodule KlassHero.Provider.Application.Commands.Providers.ProviderVerification
       params = %{provider_id: provider.id, admin_id: admin.id}
 
       # First verification
-      {:ok, verified1} = VerifyProvider.execute(params)
+      {:ok, verified1} = Provider.verify_provider(params.provider_id, params.admin_id)
       assert verified1.verified == true
 
       # Second verification should still succeed
-      {:ok, verified2} = VerifyProvider.execute(params)
+      {:ok, verified2} = Provider.verify_provider(params.provider_id, params.admin_id)
       assert verified2.verified == true
 
       # verified_at may be updated or stay the same depending on implementation
@@ -85,11 +84,11 @@ defmodule KlassHero.Provider.Application.Commands.Providers.ProviderVerification
   describe "UnverifyProvider.execute/1" do
     test "sets provider as unverified", %{provider: provider, admin: admin} do
       # First verify the provider
-      VerifyProvider.execute(%{provider_id: provider.id, admin_id: admin.id})
+      Provider.verify_provider(provider.id, admin.id)
 
       # Then unverify
       params = %{provider_id: provider.id, admin_id: admin.id}
-      assert {:ok, unverified} = UnverifyProvider.execute(params)
+      assert {:ok, unverified} = Provider.unverify_provider(params.provider_id, params.admin_id)
 
       assert unverified.verified == false
       assert unverified.verified_at == nil
@@ -97,14 +96,14 @@ defmodule KlassHero.Provider.Application.Commands.Providers.ProviderVerification
 
     test "publishes integration event", %{provider: provider, admin: admin} do
       # First verify
-      VerifyProvider.execute(%{provider_id: provider.id, admin_id: admin.id})
+      Provider.verify_provider(provider.id, admin.id)
 
       # Clear events from verify operation
       clear_integration_events()
 
       # Then unverify
       params = %{provider_id: provider.id, admin_id: admin.id}
-      {:ok, _} = UnverifyProvider.execute(params)
+      {:ok, _} = Provider.unverify_provider(params.provider_id, params.admin_id)
 
       event = assert_integration_event_published(:provider_unverified)
       assert event.entity_id == provider.id
@@ -115,7 +114,7 @@ defmodule KlassHero.Provider.Application.Commands.Providers.ProviderVerification
     test "returns error when provider not found", %{admin: admin} do
       params = %{provider_id: Ecto.UUID.generate(), admin_id: admin.id}
 
-      assert {:error, :not_found} = UnverifyProvider.execute(params)
+      assert {:error, :not_found} = Provider.unverify_provider(params.provider_id, params.admin_id)
     end
 
     test "is idempotent - unverifying already unverified provider succeeds", %{
@@ -126,11 +125,11 @@ defmodule KlassHero.Provider.Application.Commands.Providers.ProviderVerification
       params = %{provider_id: provider.id, admin_id: admin.id}
 
       # First unverify (already unverified)
-      {:ok, unverified1} = UnverifyProvider.execute(params)
+      {:ok, unverified1} = Provider.unverify_provider(params.provider_id, params.admin_id)
       assert unverified1.verified == false
 
       # Second unverify
-      {:ok, unverified2} = UnverifyProvider.execute(params)
+      {:ok, unverified2} = Provider.unverify_provider(params.provider_id, params.admin_id)
       assert unverified2.verified == false
       assert unverified2.verified_at == nil
     end
